chore: Improve error messages and feedback
