+++ packages/shared/lib/appUpdater.ts
@@ -1,9 +1,13 @@
 import { Electron, NativeProgress, VersionDetails } from 'shared/lib/electron'
 import { localize } from 'shared/lib/i18n'
 import {
-    NotificationData, NOTIFICATION_TIMEOUT_NEVER, removeDisplayNotification, showAppNotification,
-    updateDisplayNotification, updateDisplayNotificationProgress
+    NOTIFICATION_TIMEOUT_NEVER,
+    removeDisplayNotification,
+    showAppNotification,
+    updateDisplayNotification,
+    updateDisplayNotificationProgress
 } from 'shared/lib/notifications'
+import type { NotificationData } from 'shared/lib/typings/notification'
 import { writable } from 'svelte/store'
 
 const DEFAULT_APP_UPDATER_POLL_INTERVAL = 900000 // 15 Minutes
