+++ packages/shared/lib/notifications.ts
@@ -3,26 +3,10 @@ import { generateRandomId } from 'shared/lib/utils';
 import { get, writable } from 'svelte/store';
 import { Electron } from './electron';
 
+import type { NotificationData } from './typings/notification';
+
 const NOTIFICATION_TIMEOUT_DEFAULT = 5000
 export const NOTIFICATION_TIMEOUT_NEVER = -1
-
-export type NotificationAction = {
-    label: string
-    isPrimary?: boolean
-    callback?: (notificationData: NotificationData, actionIndex: number) => void
-}
-
-export type NotificationData = {
-    type: 'info' | 'warning' | 'error'
-    message: string
-    progress?: number
-    subMessage?: string
-    actions?: NotificationAction[]
-    id?: string
-    ts?: number
-    timeout?: number
-    contextData?: any
-}
 
 export const displayNotifications = writable<Array<NotificationData>>([])
