+++ packages/shared/lib/typings/index.ts
@@ -4,7 +4,8 @@ import * as bridge from './bridge'
 import * as client from './client'
 import * as events from './events'
 import * as message from './message'
+import * as notification from './notification'
 import * as ledger from './ledger'
 import * as wallet from './wallet'
 
-export { account, address, bridge, client, events, ledger, message, wallet }
+export { account, address, bridge, client, events, ledger, message, notification, wallet }
