+++ packages/shared/lib/typings/notification.ts
@@ -1,0 +1,19 @@
+export type NotificationType = 'info' | 'warning' | 'error'
+
+export type NotificationAction = {
+    label: string
+    isPrimary?: boolean
+    callback?: (notificationData: NotificationData, actionIndex: number) => void
+}
+
+export type NotificationData = {
+    type: NotificationType
+    message: string
+    progress?: number
+    subMessage?: string
+    actions?: NotificationAction[]
+    id?: string
+    ts?: number
+    timeout?: number
+    contextData?: any
+}
